Create navigation and hero sections
A new file for a regular font weight was added to the fonts folder.

To the index.html was added the link to the stylesheet also the
navigation and the hero sections were added to the body. All this
implementing the use of semantic tags for meaningful html structure.
Including the necessary content for this sections.

In the CSS file added a new font face for a regular font weight,
for the html element a font size of  62.5% was defined, this for a easy
change of font sizes using REM as unit. lastly, the styles needed for
the nav and hero sections were added.